Updated for some signed/unsigned fixes.

+++ src/sack.h
@@ -12193,7 +12193,7 @@
                   , CTEXTSTR name
                   , uintptr_t psv
                   );
- CONFIGSCR_PROC( uintptr_t, ProcessConfigurationInput )( PCONFIG_HANDLER pch, CTEXTSTR block, int size, uintptr_t psv );
+ CONFIGSCR_PROC( uintptr_t, ProcessConfigurationInput )( PCONFIG_HANDLER pch, CTEXTSTR block, size_t size, uintptr_t psv );
  /*
   * TO BE IMPLEMENTED
   *
